+++ cli/push.go
@@ -5,10 +5,14 @@ import (
 	"errors"
 	"fmt"
 	"os"
+	"regexp"
+	"strings"
 
+	v1 "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/spf13/cobra"
 	"k8s.io/kubectl/pkg/util/templates"
 
+	"github.com/uor-framework/client/builder/api/v1alpha1"
 	"github.com/uor-framework/client/builder/config"
 	"github.com/uor-framework/client/registryclient/orasclient"
 	"github.com/uor-framework/client/util/workspace"
@@ -157,4 +161,37 @@ func (o *PushOptions) Run(ctx context.Context) error {
 	o.Logger.Infof("Artifact %s published to %s\n", desc.Digest, o.Destination)
 
 	return nil
+}
+
+// AddDescriptors adds the attributes of each file listed in the config
+// to the annotations of its respective descriptor.
+func AddDescriptors(d []v1.Descriptor, c v1alpha1.DataSetConfiguration) ([]v1.Descriptor, error) {
+	// For each descriptor
+	for i1, desc := range d {
+		// Get the filename of the block
+		filename := desc.Annotations["org.opencontainers.image.title"]
+		// For each file in the config
+		for i2, file := range c.Files {
+			// If the filename of the block matches the filename of the file in the config
+			// If the config has a grouping declared, make a valid regex.
+			if strings.Contains(file.File, "*") && !strings.Contains(file.File, ".*") {
+				file.File = strings.Replace(file.File, "*", ".*", -1)
+			}
+			namesearch, err := regexp.Compile(file.File)
+			if err != nil {
+				return []v1.Descriptor{}, err
+			}
+			// Find the matching descriptor
+			if namesearch.Match([]byte(filename)) {
+				// Get the k/v pairs from the config and add them to the block's annotations.
+				for k, v := range c.Files[i2].Attributes {
+					d[i1].Annotations[k] = v
+				}
+			} else {
+				// If the block does not have a corresponding config element, skip it.
+				continue
+			}
+		}
+	}
+	return d, nil
 }
